remove TemplateCallExpression and implement additional edges in CallExpression instead
